+++ ErsatzTV.Application/Graphics/GraphicsElementViewModel.cs
@@ -1,0 +1,3 @@
+namespace ErsatzTV.Application.Graphics;
+
+public record GraphicsElementViewModel(int Id, string Name);