+++ ErsatzTV.Application/Graphics/Queries/GetAllGraphicsElements.cs
@@ -1,0 +1,3 @@
+namespace ErsatzTV.Application.Graphics;
+
+public record GetAllGraphicsElements : IRequest<List<GraphicsElementViewModel>>;