+++ ErsatzTV.Core/Domain/PlayoutItemWatermark.cs
@@ -4,6 +4,6 @@ public class PlayoutItemWatermark
 {
     public int PlayoutItemId { get; set; }
     public PlayoutItem PlayoutItem { get; set; }
-    public ChannelWatermark Watermark { get; set; }
     public int? WatermarkId { get; set; }
+    public ChannelWatermark Watermark { get; set; }
 }