+++ ErsatzTV.Core/Interfaces/Streaming/GraphicsEngineContext.cs
@@ -1,5 +1,6 @@
 using ErsatzTV.Core.Domain;
 using ErsatzTV.Core.FFmpeg;
+using ErsatzTV.Core.Graphics;
 
 namespace ErsatzTV.Core.Interfaces.Streaming;
 
@@ -15,3 +16,5 @@ public record GraphicsEngineContext(
 public abstract record GraphicsElementContext;
 
 public record WatermarkElementContext(WatermarkOptions Options) : GraphicsElementContext;
+
+public record TextElementContext(TextGraphicsElement TextElement) : GraphicsElementContext;