+++ CHANGELOG.md
@@ -19,6 +19,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - `LinearFadePoints(time, start, peakStart, peakEnd, end)`
 - Add `Z-Index` to watermark editor
   - The graphics engine will order by z-index when overlaying watermarks
+- Add *experimental* `Graphics Element` template system
+  - Graphics elements are defined in YAML files inside ETV config folder / templates / graphics-elements subfolder
+  - Add `Text` graphics element type
+    - Supported in playback troubleshooting
+    - Displays multi-line text in a specified font, color, location, z-index
+    - Supports constant opacity and opacity expression
 
 ### Fix
 - Fix database operations that were slowing down playout builds